PLAT-305: moved copypasted code to helpers
