+++ ledger-core/v2/testutils/localdebug.go
@@ -3,8 +3,8 @@
 // This material is licensed under the Insolar License version 1.0,
 // available at https://github.com/insolar/assured-ledger/blob/master/LICENSE.md.
 
-// +build localdebug
-
 package testutils
 
-const LocalDebug = true
+func IsLocalDebug() bool {
+	return LocalDebug
+}
