+++ ledger-core/v2/testutils/localdebug_default.go
@@ -8,7 +8,3 @@
 package testutils
 
 const LocalDebug = false
-
-func IsLocalDebug() bool {
-	return LocalDebug
-}
